Reorganise swarp
