+++ mirar/processors/astromatic/swarp/__init__.py
@@ -1,1 +1,7 @@
-from mirar.processors.astromatic.swarp.swarp import run_swarp
+"""
+Processors for interfacing with the Swarp software.
+"""
+from mirar.processors.astromatic.swarp.component_images import (
+    ReloadSwarpComponentImages,
+)
+from mirar.processors.astromatic.swarp.swarp import Swarp
